pyproject.toml
[tool.poetry]
name = "portal-backend"
version = "0.1.0"
authors = ["rally.lin@duke.edu"]
description = ""
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
pydantic = "^2.11.7"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic-settings = "^2.10.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
python-dotenv = "^1.1.1"
sqlalchemy = "^2.0.43"
pytest-asyncio = "^1.1.0"
boto3 = "^1.40.29"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
httpx = "^0.28.1"
alembic = "^1.16.5"
psycopg2 = "^2.9.10"
pytest-cov = "^7.0.0"
